add new fields to products and setting table and queries to fill them

+++ Modules/Setting/Database/Migrations/2021_12_15_220744_create_settings_table.php
@@ -20,7 +20,7 @@ class CreateSettingsTable extends Migration
             $table->text('keywords')->nullable();
             $table->text('logo')->nullable();
             $table->text('icon')->nullable();
-            $table->integer('low_count_products')->default(0)->comment('count of products which have low marketable number');
+            $table->smallInteger('low_count_products')->default(0)->comment('count of products which have low marketable number');
             $table->float('rating_score')->default(0)->comment('calculate avg rating of all items');
             $table->text('author')->nullable();
             $table->text('address')->nullable();
